+++ trunk/nulldc/nullDC/dc/maple/maple_if.cpp
@@ -324,8 +324,8 @@ void maple_plugins_Init()
 			if (strcmp(plugin,"NULL")!=0)
 			{
 				maple_device* plugin_dev=FindMapleDevice(plugin);
-				plugin_dev->CreateInstance(plugin_dev,MapleDevices[i][j],GetMapleAddress(i,j));
-				MapleDevices[i][j].port=GetMapleAddress(i,j);
+				plugin_dev->CreateInstance(plugin_dev,MapleDevices[i][j],(u8)GetMapleAddress(i,j));
+				MapleDevices[i][j].port=(u8)GetMapleAddress(i,j);
 				MapleDevices[i][j].Connected=true;
 			}
 			else
