+++ trunk/nulldc/nullDC/plugins/plugin_manager.cpp
@@ -403,7 +403,7 @@ void plugins_Init()
 
 void plugins_Term()
 {
-	for (int i=libMaple.size()-1;i>=0;i--)
+	for (size_t i=libMaple.size()-1;i>=0;i--)
 	{
 		libMaple[i]->info.Term(PluginType::MapleDevice);
 	}
@@ -518,7 +518,7 @@ void plugins_ThreadInit()
 
 void plugins_ThreadTerm()
 {
-	for (int i=libMaple.size()-1;i>=0;i--)
+	for (size_t i=libMaple.size()-1;i>=0;i--)
 	{
 		libMaple[i]->info.ThreadTerm(PluginType::MapleDevice);
 	}
